+++ selene/support/conditions/not_.py
@@ -22,37 +22,36 @@
 import warnings
 from typing import Any
 
-from selene import match
-
+from selene.core import match as _match
 
 # --- be.* conditions --- #
 
 
-visible = match.element_is_visible.not_
-hidden = match.element_is_hidden.not_
+visible = _match.element_is_visible.not_
+hidden = _match.element_is_hidden.not_
 
-present = match.element_is_present.not_
-in_dom = match.element_is_present.not_    # todo: do we need both present and in_dom?
-existing = match.element_is_present.not_  # todo: consider deprecating
+present = _match.element_is_present.not_
+in_dom = _match.element_is_present.not_    # todo: do we need both present and in_dom?
+existing = _match.element_is_present.not_  # todo: consider deprecating
 
-absent = match.element_is_absent.not_
+absent = _match.element_is_absent.not_
 
-enabled = match.element_is_enabled.not_
-disabled = match.element_is_disabled.not_
+enabled = _match.element_is_enabled.not_
+disabled = _match.element_is_disabled.not_
 
-blank = match.element_is_blank.not_
+blank = _match.element_is_blank.not_
 
 
 # --- have.* conditions --- #
 
 
-def exact_text(value) -> match.ElementCondition:
-    return match.element_has_exact_text(value).not_
+def exact_text(value) -> _match.ElementCondition:
+    return _match.element_has_exact_text(value).not_
 
 
 # todo: consider accepting int
-def text(partial_value) -> match.ElementCondition:
-    return match.element_has_text(partial_value).not_
+def text(partial_value) -> _match.ElementCondition:
+    return _match.element_has_text(partial_value).not_
 
 
 def attribute(name: str, value: str = None):
@@ -60,21 +59,21 @@ def attribute(name: str, value: str = None):
         warnings.warn(
             'passing second argument is deprecated; use have.attribute(foo).value(bar) instead',
             DeprecationWarning)
-        return match.element_has_attribute(name).value(value).not_
+        return _match.element_has_attribute(name).value(value).not_
 
-    original = match.element_has_attribute(name)
+    original = _match.element_has_attribute(name)
     negated = original.not_
 
-    def value(self, expected: str) -> match.ElementCondition:
+    def value(self, expected: str) -> _match.ElementCondition:
         return original.value(expected).not_
 
-    def value_containing(self, expected: str) -> match.ElementCondition:
+    def value_containing(self, expected: str) -> _match.ElementCondition:
         return original.value_containing(expected).not_
 
-    def values(self, *expected: str) -> match.CollectionCondition:
+    def values(self, *expected: str) -> _match.CollectionCondition:
         return original.values(*expected).not_
 
-    def values_containing(self, *expected: str) -> match.CollectionCondition:
+    def values_containing(self, *expected: str) -> _match.CollectionCondition:
         return original.values_containing(*expected).not_
 
     negated.value = value
@@ -90,21 +89,21 @@ def js_property(name: str, value: str = None):
         warnings.warn(
             'passing second argument is deprecated; use have.js_property(foo).value(bar) instead',
             DeprecationWarning)
-        return match.element_has_js_property(name).value(value).not_
+        return _match.element_has_js_property(name).value(value).not_
 
-    original = match.element_has_js_property(name)
+    original = _match.element_has_js_property(name)
     negated = original.not_
 
-    def value(self, expected: str) -> match.ElementCondition:
+    def value(self, expected: str) -> _match.ElementCondition:
         return original.value(expected).not_
 
-    def value_containing(self, expected: str) -> match.ElementCondition:
+    def value_containing(self, expected: str) -> _match.ElementCondition:
         return original.value_containing(expected).not_
 
-    def values(self, *expected: str) -> match.CollectionCondition:
+    def values(self, *expected: str) -> _match.CollectionCondition:
         return original.values(*expected).not_
 
-    def values_containing(self, *expected: str) -> match.CollectionCondition:
+    def values_containing(self, *expected: str) -> _match.CollectionCondition:
         return original.values_containing(*expected).not_
 
     negated.value = value
@@ -120,21 +119,21 @@ def css_property(name: str, value: str = None):
         warnings.warn(
             'passing second argument is deprecated; use have.css_property(foo).value(bar) instead',
             DeprecationWarning)
-        return match.element_has_css_property(name).value(value).not_
+        return _match.element_has_css_property(name).value(value).not_
 
-    original = match.element_has_css_property(name)
+    original = _match.element_has_css_property(name)
     negated = original.not_
 
-    def value(self, expected: str) -> match.ElementCondition:
+    def value(self, expected: str) -> _match.ElementCondition:
         return original.value(expected).not_
 
-    def value_containing(self, expected: str) -> match.ElementCondition:
+    def value_containing(self, expected: str) -> _match.ElementCondition:
         return original.value_containing(expected).not_
 
-    def values(self, *expected: str) -> match.CollectionCondition:
+    def values(self, *expected: str) -> _match.CollectionCondition:
         return original.values(*expected).not_
 
-    def values_containing(self, *expected: str) -> match.CollectionCondition:
+    def values_containing(self, *expected: str) -> _match.CollectionCondition:
         return original.values_containing(*expected).not_
 
     negated.value = value
@@ -145,103 +144,103 @@ def css_property(name: str, value: str = None):
     return negated
 
 
-def value(text) -> match.ElementCondition:
-    return match.element_has_value(text).not_
+def value(text) -> _match.ElementCondition:
+    return _match.element_has_value(text).not_
 
 
-def value_containing(partial_text) -> match.ElementCondition:
-    return match.element_has_value_containing(partial_text).not_
+def value_containing(partial_text) -> _match.ElementCondition:
+    return _match.element_has_value_containing(partial_text).not_
 
 
-def css_class(name) -> match.ElementCondition:
-    return match.element_has_css_class(name).not_
+def css_class(name) -> _match.ElementCondition:
+    return _match.element_has_css_class(name).not_
 
 
-def tag(name: str) -> match.ElementCondition:
-    return match.element_has_tag(name).not_
+def tag(name: str) -> _match.ElementCondition:
+    return _match.element_has_tag(name).not_
 
 
-def tag_containing(name: str) -> match.ElementCondition:
-    return match.element_has_tag_containing(name).not_
+def tag_containing(name: str) -> _match.ElementCondition:
+    return _match.element_has_tag_containing(name).not_
 
 
 # *** SeleneCollection conditions ***
 
 
-def size(number: int) -> match.CollectionCondition:
-    return match.collection_has_size(number).not_
+def size(number: int) -> _match.CollectionCondition:
+    return _match.collection_has_size(number).not_
 
 
-def size_less_than(number: int) -> match.CollectionCondition:
-    return match.collection_has_size_less_than(number).not_
+def size_less_than(number: int) -> _match.CollectionCondition:
+    return _match.collection_has_size_less_than(number).not_
 
 
-def size_less_than_or_equal(number: int) -> match.CollectionCondition:
-    return match.collection_has_size_less_than_or_equal(number).not_
+def size_less_than_or_equal(number: int) -> _match.CollectionCondition:
+    return _match.collection_has_size_less_than_or_equal(number).not_
 
 
-def size_greater_than(number: int) -> match.CollectionCondition:
-    return match.collection_has_size_greater_than(number).not_
+def size_greater_than(number: int) -> _match.CollectionCondition:
+    return _match.collection_has_size_greater_than(number).not_
 
 
-def size_at_least(number: int) -> match.CollectionCondition:
+def size_at_least(number: int) -> _match.CollectionCondition:
     warnings.warn('might be deprecated; use have.size_greater_than_or_equal instead', PendingDeprecationWarning)
-    return match.collection_has_size_greater_than_or_equal(number).not_
+    return _match.collection_has_size_greater_than_or_equal(number).not_
 
 
-def size_greater_than_or_equal(number: int) -> match.CollectionCondition:
-    return match.collection_has_size_greater_than_or_equal(number).not_
+def size_greater_than_or_equal(number: int) -> _match.CollectionCondition:
+    return _match.collection_has_size_greater_than_or_equal(number).not_
 
 
 # todo: consider accepting ints
-def texts(*partial_values: str) -> match.CollectionCondition:
-    return match.collection_has_texts(*partial_values).not_
+def texts(*partial_values: str) -> _match.CollectionCondition:
+    return _match.collection_has_texts(*partial_values).not_
 
 
-def exact_texts(*values: str) -> match.CollectionCondition:
-    return match.collection_has_exact_texts(*values).not_
+def exact_texts(*values: str) -> _match.CollectionCondition:
+    return _match.collection_has_exact_texts(*values).not_
 
 
-def url(exact_value: str) -> match.BrowserCondition:
-    return match.browser_has_url(exact_value).not_
+def url(exact_value: str) -> _match.BrowserCondition:
+    return _match.browser_has_url(exact_value).not_
 
 
-def url_containing(partial_value: str) -> match.BrowserCondition:
-    return match.browser_has_url_containing(partial_value).not_
+def url_containing(partial_value: str) -> _match.BrowserCondition:
+    return _match.browser_has_url_containing(partial_value).not_
 
 
-def title(exact_value: str) -> match.BrowserCondition:
-    return match.browser_has_title(exact_value).not_
+def title(exact_value: str) -> _match.BrowserCondition:
+    return _match.browser_has_title(exact_value).not_
 
 
-def title_containing(partial_value: str) -> match.BrowserCondition:
-    return match.browser_has_title_containing(partial_value).not_
+def title_containing(partial_value: str) -> _match.BrowserCondition:
+    return _match.browser_has_title_containing(partial_value).not_
 
 
-def tabs_number(value: int) -> match.BrowserCondition:
-    return match.browser_has_tabs_number(value).not_
+def tabs_number(value: int) -> _match.BrowserCondition:
+    return _match.browser_has_tabs_number(value).not_
 
 
-def tabs_number_less_than(value: int) -> match.BrowserCondition:
-    return match.browser_has_tabs_number_less_than(value).not_
+def tabs_number_less_than(value: int) -> _match.BrowserCondition:
+    return _match.browser_has_tabs_number_less_than(value).not_
 
 
-def tabs_number_less_than_or_equal(value: int) -> match.BrowserCondition:
-    return match.browser_has_tabs_number_less_than_or_equal(value).not_
+def tabs_number_less_than_or_equal(value: int) -> _match.BrowserCondition:
+    return _match.browser_has_tabs_number_less_than_or_equal(value).not_
 
 
-def tabs_number_greater_than(value: int) -> match.BrowserCondition:
-    return match.browser_has_tabs_number_greater_than(value).not_
+def tabs_number_greater_than(value: int) -> _match.BrowserCondition:
+    return _match.browser_has_tabs_number_greater_than(value).not_
 
 
-def tabs_number_greater_than_or_equal(value: int) -> match.BrowserCondition:
-    return match.browser_has_tabs_number_greater_than_or_equal(value).not_
+def tabs_number_greater_than_or_equal(value: int) -> _match.BrowserCondition:
+    return _match.browser_has_tabs_number_greater_than_or_equal(value).not_
 
 
-def js_returned_true(script_to_return_bool: str) -> match.BrowserCondition:
+def js_returned_true(script_to_return_bool: str) -> _match.BrowserCondition:
     warnings.warn('might be deprecated; use have.js_returned(True, ...) instead', PendingDeprecationWarning)
-    return match.browser_has_js_returned(True, script_to_return_bool).not_
+    return _match.browser_has_js_returned(True, script_to_return_bool).not_
 
 
-def js_returned(expected: Any, script: str, *args) -> match.BrowserCondition:
-    return match.browser_has_js_returned(expected, script, *args).not_
+def js_returned(expected: Any, script: str, *args) -> _match.BrowserCondition:
+    return _match.browser_has_js_returned(expected, script, *args).not_