+++ selene/support/shared/__init__.py
@@ -23,7 +23,6 @@
 
 from __future__ import annotations
 
-from selene.configuration import Config
 from selene.support.shared.browser import SharedBrowser
 from selene.support.shared.config import SharedConfig